rename of class
renamed pimon to PiWeM since that is what it is.

+++ PiWeM/PiWeM.py
@@ -6,8 +6,10 @@ from PIL import ImageDraw
 from PIL import ImageFont
 from dht11 import dht11
 
-class PIMON:
+class PIWEM:
     loop_int = 0
+    take_picture_flag = 1
+    verbose = 0
     pin = 0
     brightness = 0
     resolution = (0,0)
@@ -69,6 +71,10 @@ class PIMON:
         return 0
 
 
+    def get_berometer_data(self):
+
+        return 0
+
     def take_picture(self):
         print("Take Picture.")
         image_name = self.tmp_dir + "Image_" + str(self.loop_int) + ".jpg"
@@ -76,22 +82,33 @@ class PIMON:
         return image_name
 
 
-    def write_text_to_image(self):
-        image = self.take_picture()
-        img = Image.open( image)
-        filename = os.path.basename(image)
-        draw = ImageDraw.Draw(img)
-        #font = ImageFont.truetype(<font-file>, <font-size>)
-        font = ImageFont.truetype(self.text_font, 40)
+    def get_data_trigger(self):
+        if self.take_picture_flag:
+            image = self.take_picture()
+            img = Image.open( image)
+            filename = os.path.basename(image)
+            draw = ImageDraw.Draw(img)
+            #font = ImageFont.truetype(<font-file>, <font-size>)
+            font = ImageFont.truetype(self.text_font, 40)
+
         data = self.get_temp_humidity_data()
         if data:
             self.insert_data(data)
-            #draw.text((x, y),"Sample Text",(r,g,b))
-            draw.text((10, 0), "Humidity:       " + str(data[0]) + "%" ,self.text_color, font=font)
-            draw.text((10, 60), "Temperature: " + str(data[1]) + "C / " + str(data[2]) + "F" ,self.text_color, font=font)
-            img.save(self.output_path + filename, quality=self.jpeg_quality)
-            os.remove(image)
-            print("Saved Image: " + self.output_path + filename)
+            if self.verbose:
+                ts = time.time()
+                timestamp = datetime.datetime.fromtimestamp(ts).strftime('%Y-%m-%d %H:%M:%S')
+
+                print("TimeStamp: " + timestamp)
+                print("Humidity:       " + str(data[0]) + "%")
+                print("Temperature: " + str(data[1]) + "C / " + str(data[2]) + "F")
+
+            if self.take_picture_flag:
+                #draw.text((x, y),"Sample Text",(r,g,b))
+                draw.text((10, 0), "Humidity:       " + str(data[0]) + "%" ,self.text_color, font=font)
+                draw.text((10, 60), "Temperature: " + str(data[1]) + "C / " + str(data[2]) + "F" ,self.text_color, font=font)
+                img.save(self.output_path + filename, quality=self.jpeg_quality)
+                os.remove(image)
+                print("Saved Image: " + self.output_path + filename)
             return 1
         else:
             print("Data return error.")
